Fixed Quoted and UI issues.
