+++ src/main/java/de/blau/android/bookmarks/BookmarkIO.java
@@ -3,29 +3,36 @@ package de.blau.android.bookmarks;
 import android.content.Context;
 
 import java.util.ArrayList;
+import java.util.List;
 
 import de.blau.android.osm.ViewBox;
 import de.blau.android.util.SavingHelper;
 
 
 public class  BookmarkIO {
-    private BookmarksStorage currentBookmarkStorage = new BookmarksStorage();
-    private ArrayList<BookmarksStorage> bookmarksStorage = new ArrayList<>();
+    private BookmarksStorage currentBookmarkStorage;
+    private List<BookmarksStorage> bookmarksStorage;
     private final String fileName = "bookmarks.ser";
-    SavingHelper<ArrayList<BookmarksStorage>> savingHelper = new SavingHelper();
+    SavingHelper<ArrayList<BookmarksStorage>> savingHelper;
 
+
+    public BookmarkIO(){
+        this.currentBookmarkStorage = new BookmarksStorage();
+        this.bookmarksStorage = new ArrayList<>();
+        this.savingHelper = new SavingHelper<>();
+    }
     /**
      * adding a bookmark storage object to arraylist
      */
-    public void addDatatolist(String s, ViewBox viewbox) {
-        currentBookmarkStorage.set(s, viewbox);
+    public void addDatatolist(String comment, ViewBox viewbox) {
+        currentBookmarkStorage.set(comment, viewbox);
         bookmarksStorage.add(currentBookmarkStorage);
     }
     /**
      * (over)Writes the bookmark file to storage.
      */
-    public void writeList(Context context) {
-        savingHelper.save(context, fileName, bookmarksStorage, true);
+    public void writeList(Context context,ArrayList<BookmarksStorage> bookmarksStorage) {
+        savingHelper.save(context, fileName, (ArrayList<BookmarksStorage>) bookmarksStorage, true);
     }
     /**
      * Reads the bookmark file from storage.
@@ -34,11 +41,9 @@ public class  BookmarkIO {
 
         ArrayList<BookmarksStorage> savedList = savingHelper.load(context,fileName,true);
         if(savedList==null){
-            writeList(context);
-            readList(context);
+            return (ArrayList<BookmarksStorage>) this.bookmarksStorage;
         }
         return savedList;
-
     }
     /**
      * Utility for bookmark saver
@@ -46,9 +51,7 @@ public class  BookmarkIO {
     public void writer(Context context,String message,ViewBox viewBox){
         this.bookmarksStorage = readList(context);
         addDatatolist(message,viewBox);
-        writeList(context);
-
-
+        writeList(context, (ArrayList<BookmarksStorage>) this.bookmarksStorage);
     }
 
 }
