+++ src/main/java/de/blau/android/dialogs/BookmarksHandler.java
@@ -1,11 +1,10 @@
-package de.blau.android.bookmarks;
+package de.blau.android.dialogs;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatDialog;
 import androidx.fragment.app.FragmentActivity;
 
 import de.blau.android.R;
-import de.blau.android.dialogs.TextLineDialog;
 
 
 public class BookmarksHandler {
@@ -20,7 +19,7 @@ public class BookmarksHandler {
 
     public static void get(@NonNull final FragmentActivity activity, @NonNull final BookmarksHandler.HandleResult handler){
 
-        dialog = TextLineDialog.get(activity, R.string.add_to_bookmarks,R.string.bookmarks_desc,(input , check) -> {
+        dialog = TextLineDialog.get(activity, R.string.add_bookmark,R.string.bookmarks_desc,(input , check) -> {
             String comments = input.getText().toString();
             handler.onSuccess(comments,activity);
             dismiss();
