+++ src/main/res/menu/bookmark_popup.xml
@@ -1,12 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
-        xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/show"
-        android:title="@string/gotobookmark"
-         />
     <item
         android:id="@+id/bookmarkdiscard"
         android:title="@string/discard"
-                                />
+        />
 </menu>
